chore(server): Set node version to 8 for Travis CI
Yarn is failing with node 9 because graphql-js v14.0.0 removed support for node 9 in the "engines" field of package.json.

Looks like this will be sorted out in the next release (https://github.com/graphql/graphql-js/pull/1508) but for now it is easiest to use node v8 to make the build pass.