feat(core): Use batching when reindexing search index
On a test dataset of ~9k ProductVariants, the reindex was using an extra ~550MB. With this batching, the memory usage stays pretty flat around the idle level. The trade-off is that the operation takes around 20% longer, but during that time the server is still responsive (albeit with increased latency). Formerly, the server would not serve requests for the duration of the reindexing op.