fix(core): Fix order calculation with over 1000 active Promotions
We were previously getting the promotions via the ListQueryBuilder,
which has a default limit of 1000 items. So a user who was doing a data
import which was creating over 1000 Promotions found that, after that limit
was reached, the expected discounts were no longer being applied to the imported
orders.

This commit fixes it by making a direct DB query with no upper limit.